requirements.txt
asgiref==3.5.2
astroid==2.12.10
autopep8==1.7.0
dill==0.3.5.1
disposable-email-domains==0.0.81
Django==4.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
PyJWT==2.5.0
pylint==2.15.3
pytz==2022.2.1
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
